drf_requirements.txt
anyascii==0.3.1
asgiref==3.6.0
beautifulsoup4==4.9.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
coderedcms==1.0.1
colorama==0.4.6
cryptography==39.0.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.0.9
django-allauth==0.52.0
django-bootstrap==0.2.4
django-bootstrap-icons==0.8.2
django-bootstrap5==21.3
django-eventtools==1.0.3
django-filter==21.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-sendmail-backend==0.1.2
django-taggit==2.1.0
django-treebeard==4.6.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filelock==3.9.0
future==0.18.3
geocoder==1.38.1
html5lib==1.1
icalendar==4.1.0
idna==3.4
l18n==2021.3
oauthlib==3.2.2
openpyxl==3.1.0
Pillow==9.4.0
platformdirs==2.6.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
ratelim==0.1.6
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tablib==3.3.0
telepath==0.3
tzdata==2022.7
urllib3==1.26.14
virtualenv==20.17.1
wagtail==3.0.3
wagtail-cache==2.2.0
wagtail-seo==2.3.0
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.8
xlwt==1.3.0
